chore: move browser tests to shared platform tests
`externalEntropy.test.ts`, the database migration test, and a logger
test are browser specific, so we don't move these.
