+++ crypto-ffi/bindings/js/shared/test/errors.test.ts
